+++ ThermalConsole/Program.cs
@@ -35,7 +35,7 @@ namespace ThermalConsole
         static void Main(string[] args)
         {
             const string phoenixPort = "COM1";
-           const string reliancePort = "COM23";            
+           const string reliancePort = "COM12";            
 
             const int captureRate = 10; // number of seconds between capture
             
@@ -84,7 +84,7 @@ namespace ThermalConsole
             var document = new StandardDocument()
             {
                 // Don't forget to set your codepage!
-                CodePage = CodePages.ASCII,
+                CodePage = CodePages.CPSPACE,
             };
 
             document.Sections.Add(header);
@@ -117,7 +117,7 @@ namespace ThermalConsole
                     header.Content = string.Format("Capture #{0} (ЫВФАЫВМОЫВАП)", count);
 
                     // Printer the timestamp document
-                    timestamp.Content = string.Format("{1}", count++, now);
+                    timestamp.Content = string.Format("{0}: {1}", count++, now);
 
                     // Get the latest printer status. Note that reliance and phoenix have
                     // slightly different args to this get status command